Fix up build on Solaris with data_packaging=library
The problem with the build was that the library directories were added
to the incorrect library path. The code uses LD_LIBRARY_PATH by default,
but on our Solaris machines, we want LD_LIBRARY_PATH_64.

There is a Make variable LDLIBRARYPATH_ENVVAR that controls the variable
name. Set that variable on Solaris 64-bit to use LD_LIBRARY_PATH_64.

In the process, generalized the way arguments are presented to the make
command.